IGNITE-25747 Configuration compatibility. Support annotated values validation.
Review.
