+++ modules/runner/src/test/java/org/apache/ignite/internal/configuration/compatibility/framework/ConfigAnnotationsValidator.java
@@ -17,41 +17,48 @@
 
 package org.apache.ignite.internal.configuration.compatibility.framework;
 
-import static org.apache.ignite.internal.lang.IgniteStringFormatter.format;
-
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.TreeSet;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 import org.apache.ignite.configuration.validation.ExceptKeys;
 import org.apache.ignite.configuration.validation.OneOf;
 import org.apache.ignite.configuration.validation.Range;
+import org.apache.ignite.internal.configuration.compatibility.framework.annotations.DefaultAnnotationCompatibilityValidator;
+import org.apache.ignite.internal.configuration.compatibility.framework.annotations.ExceptKeysAnnotationCompatibilityValidator;
+import org.apache.ignite.internal.configuration.compatibility.framework.annotations.OneOfAnnotationCompatibilityValidator;
+import org.apache.ignite.internal.configuration.compatibility.framework.annotations.RangeAnnotationCompatibilityValidator;
 
 /**
  * Configuration annotation validator.
  */
 final class ConfigAnnotationsValidator {
 
-    private final Map<String, AnnotationValidator> validators;
+    private static final DefaultAnnotationCompatibilityValidator DEFAULT_VALIDATOR = new DefaultAnnotationCompatibilityValidator();
 
+    private final Map<String, AnnotationCompatibilityValidator> validators;
+
+    /** Constructor. */
     ConfigAnnotationsValidator() {
         this(Map.of(
-                Range.class.getName(), ConfigAnnotationsValidator::validateRange,
-                ExceptKeys.class.getName(), ConfigAnnotationsValidator::validateExceptKeys,
-                OneOf.class.getName(), ConfigAnnotationsValidator::validateOneOf
+                Range.class.getName(), new RangeAnnotationCompatibilityValidator(),
+                ExceptKeys.class.getName(), new ExceptKeysAnnotationCompatibilityValidator(),
+                OneOf.class.getName(), new OneOfAnnotationCompatibilityValidator()
         ));
     }
 
-    ConfigAnnotationsValidator(Map<String, AnnotationValidator> validators) {
+    /** Constructor. */
+    ConfigAnnotationsValidator(Map<String, AnnotationCompatibilityValidator> validators) {
         this.validators = Map.copyOf(validators);
     }
 
+    /**
+     * Validates annotations.
+     */
     void validate(ConfigNode candidate, ConfigNode node, List<String> errors) {
-
         List<ConfigAnnotation> currentAnnotations = node.annotations();
         List<ConfigAnnotation> candidateAnnotations = candidate.annotations();
 
@@ -110,11 +117,11 @@ final class ConfigAnnotationsValidator {
         }
 
         Set<String> changedTypes = new TreeSet<>();
-        Set<String> changedValues = new TreeSet<>();
 
         for (Map.Entry<String, ConfigAnnotationValue> entry : candidate.properties().entrySet()) {
             ConfigAnnotationValue currentValue = current.properties().get(entry.getKey());
             if (currentValue == null) {
+                // Already an error, just skip it.
                 continue;
             }
 
@@ -122,8 +129,6 @@ final class ConfigAnnotationsValidator {
 
             if (!Objects.equals(candidateValue.typeName(), currentValue.typeName())) {
                 changedTypes.add(entry.getKey());
-            } else if (!Objects.equals(candidateValue, currentValue)) {
-                changedValues.add(entry.getKey());
             }
         }
 
@@ -131,98 +136,13 @@ final class ConfigAnnotationsValidator {
             errors.add(candidate.name() + " properties with changed types " + changedTypes);
         }
 
-        AnnotationValidator validator = validators.get(candidate.name());
-
-        // If a specific validator is present, use it.
-        // Otherwise, expect all properties to be equal between versions.
-        if (validator != null) {
-            validator.validate(candidate, current, errors);
-        } else if (!changedValues.isEmpty()) {
-            errors.add(current.name() + " changed values " + changedValues);
-        }
+        validateSpecificAnnotation(candidate, current, errors);
     }
 
-    void validateSpecificAnnotation(
-            ConfigAnnotation candidate,
-            ConfigAnnotation current,
-            List<String> errors
-    ) {
-        AnnotationValidator validator = validators.get(candidate.name());
+    void validateSpecificAnnotation(ConfigAnnotation candidate, ConfigAnnotation current, List<String> errors) {
+        AnnotationCompatibilityValidator validator = validators.getOrDefault(candidate.name(), DEFAULT_VALIDATOR);
         assert candidate.name().equals(current.name()) : "Annotation name mismatch";
 
-        if (validator != null) {
-            validator.validate(candidate, current, errors);
-        }
-    }
-
-    @FunctionalInterface
-    interface AnnotationValidator {
-        void validate(ConfigAnnotation candidate, ConfigAnnotation current, List<String> errors);
-
-        /** Reads a value of the given annotation property. */
-        static <T> T getValue(ConfigAnnotation annotation, String name, Function<ConfigAnnotationValue, T> parse) {
-            ConfigAnnotationValue value = annotation.properties().get(name);
-            try {
-                return parse.apply(value);
-            } catch (Exception e) {
-                String errorMessage = format("Unable to read annotation property. Property: {}, annotation: {}", name, annotation.name());
-                throw new IllegalStateException(errorMessage, e);
-            }
-        }
-    }
-
-    private static void validateRange(ConfigAnnotation candidate, ConfigAnnotation current, List<String> errors) {
-        long candidateMin = AnnotationValidator.getValue(candidate, "min", (v) -> (Long) v.value());
-        long candidateMax = AnnotationValidator.getValue(candidate, "max", (v) -> (Long) v.value());
-
-        long currentMin = AnnotationValidator.getValue(current, "min", (v) -> (Long) v.value());
-        long currentMax = AnnotationValidator.getValue(current, "max", (v) -> (Long) v.value());
-
-        if (currentMin > candidateMin) {
-            errors.add("Range: min changed from " + candidateMin + " to " + currentMin);
-        }
-
-        if (currentMax < candidateMax) {
-            errors.add("Range: max changed from " + candidateMax + " to " + currentMax);
-        }
-    }
-
-    private static void validateExceptKeys(ConfigAnnotation candidate, ConfigAnnotation current, List<String> errors) {
-        List<String> candidateKeys = AnnotationValidator.getValue(candidate, "value", (v) -> (List<String>) v.value());
-        List<String> currentKeys = AnnotationValidator.getValue(current, "value", (v) -> (List<String>) v.value());
-
-        TreeSet<String> newKeys = new TreeSet<>();
-        for (String key : currentKeys) {
-            if (!candidateKeys.contains(key)) {
-                newKeys.add(key);
-            }
-        }
-
-        if (!newKeys.isEmpty()) {
-            errors.add("ExceptKeys: changed keys from " + candidateKeys + " to " + currentKeys);
-        }
-    }
-
-    private static void validateOneOf(ConfigAnnotation candidate, ConfigAnnotation current, List<String> errors) {
-        List<String> candidateKeys = AnnotationValidator.getValue(candidate, "value", (v) -> (List<String>) v.value());
-        List<String> currentKeys = AnnotationValidator.getValue(current, "value", (v) -> (List<String>) v.value());
-
-        TreeSet<String> removedKeys = new TreeSet<>();
-        for (String key : candidateKeys) {
-            if (!currentKeys.contains(key)) {
-                removedKeys.add(key);
-            }
-        }
-
-        if (!removedKeys.isEmpty()) {
-            errors.add("OneOf: changed keys from " + candidateKeys + " to " + currentKeys);
-        }
-
-        boolean candidateCaseSensitive = AnnotationValidator.getValue(candidate, "caseSensitive", (v) -> (Boolean) v.value());
-        boolean currentCaseSensitive = AnnotationValidator.getValue(current, "caseSensitive", (v) -> (Boolean) v.value());
-
-        if (candidateCaseSensitive && !currentCaseSensitive) {
-            errors.add("OneOf: keys made case insensitive");
-        }
+        validator.validate(candidate, current, errors);
     }
 }
