+++ modules/runner/src/test/java/org/apache/ignite/internal/configuration/compatibility/framework/annotations/BaseAnnotationCompatibilityValidatorSelfTest.java
@@ -1,0 +1,48 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.ignite.internal.configuration.compatibility.framework.annotations;
+
+import static org.apache.ignite.internal.lang.IgniteStringFormatter.format;
+
+import java.lang.annotation.Annotation;
+import java.lang.reflect.Field;
+import org.apache.ignite.internal.configuration.compatibility.framework.ConfigAnnotation;
+import org.apache.ignite.internal.configuration.compatibility.framework.ConfigurationTreeScanner;
+import org.apache.ignite.internal.testframework.BaseIgniteAbstractTest;
+
+/**
+ * Annotation compatibility validator tests.
+ */
+abstract class BaseAnnotationCompatibilityValidatorSelfTest extends BaseIgniteAbstractTest {
+    protected static ConfigAnnotation getAnnotation(Class<?> clazz,
+            String field,
+            String annotationName,
+            Class<? extends Annotation> annotationClass
+    ) {
+        try {
+            Field f = clazz.getField(field);
+            Annotation annotation = f.getAnnotation(annotationClass);
+            if (annotation == null) {
+                throw new IllegalStateException(format("No annotation: {} found. Class: {}, field: {}", annotationClass, clazz, field));
+            }
+            return ConfigurationTreeScanner.extractAnnotation(annotationName, annotation);
+        } catch (NoSuchFieldException e) {
+            throw new IllegalStateException(e);
+        }
+    }
+}
